Change the script to backup everything
Before, the script only copied the .txt files since I wanted to test it
on one file type. Now, I have changed it so that it will backup all
folders and files inside the directory in which the script was called
from. Also, Ive removed from config.txt any unwanted values.

Finally, I've also tidied the script even more since I saw that the for
loop was unnecessary, as with the first echo message.